Keep API auth in the endpoint, ahead of body parsing
Moving the agent secret check into a router pipeline meant Plug.Parsers
ran before it, so unauthenticated requests had their body parsed before
being rejected. Restore BobWeb.Plugs.Secret in the endpoint and instead
exempt the two public, read-only search routes from it, keeping the
pre-parse rejection for all authenticated /api routes and restoring the
test that asserts it.
